// src/enums.ts
export enum enumColorFavorito {
    azul = "azul",
    verde = "verde",
    rojo = "rojo",
    amarillo = "amarillo",
    negro = "negro",
    blanco = "blanco",
    morado = "morado",
    naranja = "naranja",
    rosa = "rosa"
}

export enum enumSexo {
    mujer = "mujer",
    hombre = "hombre",
    neutro = "neutro"
}

export enum enumTipoMail {
    personal = "personal",
    empresa = "empresa"
}

export enum enumTipoTelefono {
    movil = "movil",
    fijo = "fijo",
    empresa = "empresa"
}